+++ R/map_props_points_sera.R
@@ -163,6 +163,8 @@ srGroups <- function(map) {
 #' Getting and setting sera sequence information
 #'
 #' @param map The acmap data object
+#' @param missing_value Character to use to fill in portions of the sequence matrix
+#'   where sequence data is missing.
 #' @param value A character matrix of sequences with rows equal to the number of
 #'   sera
 #'
@@ -172,13 +174,13 @@ srGroups <- function(map) {
 
 #' @rdname srSequences
 #' @export
-srSequences <- function(map) {
+srSequences <- function(map, missing_value = "-") {
   check.acmap(map)
-  do.call(
-    rbind,
+  rbind_list_to_matrix(
     lapply(map$sera, function(sr) {
       strsplit(sr$sequence, "")[[1]]
-    })
+    }),
+    missing_value
   )
 }